alpha: change alpha variable from float to double
The float value caused the opacity to be calculated as 0 due to
precision loss when multiplying with unsigned int 0xffffffff.

+++ config.def.h
@@ -86,7 +86,7 @@ ResourcePref resources[] = {
 
 #if ALPHA_PATCH
 /* lock screen opacity */
-static const float alpha = 0.9;
+static const double alpha = 0.9;
 #endif // ALPHA_PATCH
 
 /* treat a cleared input like a wrong password (color) */
